\name{open.article}
\alias{open.article}
\title{open.article}
\description{
Open article from a saved search in browser.
}
\usage{
open.article(search = NULL, results = NULL, index = 'article.titles')
}
\arguments{
\item{search}{a string or regular expression which identifies paper to open in search results}
\item{results}{saved search results from a search.pubmed( ) call}
\item{index}{results column to search in saved search results}
}
